diffraction: extract peak plotting out of BraggPeaksPolymer
Second extraction, same pattern as the calibration one. New
`peak_visualization.py` holds the seven plotting functions whose only tie to the
class was reading one or two attributes: the radial intensity and count
profiles, the per-position count and histogram maps, the interactive pattern
browser, figure export, and the pattern grid. They take the arrays and peak
vectors they draw, so they work for any 4D-STEM dataset.

All seven are public API and three appear in 51 notebooks, so each keeps a
forwarding method on the class. The wrappers take *args/**kwargs and prepend the
data, which preserves positional calls -- verified that the notebooks'
`bp.plot_peak_count_map(peak_windows, return_values=True)` still lands
`peak_windows` on `q_ranges`.

Five shared module-level helpers moved too. The three display helpers
(`_normalized_dp`, `_resolve_intensity_map`, `_intensity_display_limits`, plus
`_mean_intensity_map`) live with the plots and are imported back for the three
peak-overlay methods that stay. The two ragged-Vector field accessors went to a
new `vector_fields.py` instead: they are needed by `estimate_peak_windows` and
`make_orientation_histogram`, which are not visualization, and importing them
from a plotting module would have been the wrong direction -- or a cycle.

Found and fixed a bug the extraction introduced: threading
`self.peak_coordinates_cartesian` in as `peaks` turned
`peaks = self.peak_coordinates_cartesian[i, j].array` into `peaks = peaks[i, j]
.array`, shadowing the parameter and failing on the second scan position. Only
two of the seven functions had any test coverage -- including none of the three
used by 51 notebooks -- so this was invisible. Added
`test_peak_visualization.py` covering the five untested ones through both entry
points, which is what caught it, and AST-scanned the rest for the same
shadowing pattern (no others).

Also repaired `test_bragg_plotting_and_save_smoke`, which monkeypatched
`interactive_output` on `bragg_peaks` only; `plot_interactive_image_map` now
resolves it from `peak_visualization`, so the patch had silently stopped
applying. It now patches both modules.

BraggPeaksPolymer: 5691 -> 4812 lines (79 -> 74 methods across both steps, since
the public plotting entry points remain as wrappers). bragg_peaks.py 7103 ->
6162, down from 8206 at the start.

Tests: 495 passed, 10 skipped.
